[refactor] Game을 수행하는 클래스와 UI를 분리
